+++ src/frontend/components/Button/Button.tsx
@@ -1,11 +1,12 @@
 import styled from 'styled-components';
 
-import { styledComponentWithProps } from '../../utils/styledComponentsTs';
 import { colorName, colors } from '../../utils/theme/theme';
 
-export const Button = styledComponentWithProps<{ variant: colorName }>(
-  styled.button,
-)`
+interface Props {
+  variant: colorName;
+}
+
+export const Button = styled.button`
   display: inline-block;
   text-align: center;
   white-space: nowrap;
@@ -19,7 +20,7 @@ export const Button = styledComponentWithProps<{ variant: colorName }>(
   transition: color 0.15s ease-in-out, background-color 0.15s ease-in-out,
     border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
 
-  ${props => `
+  ${(props: Props) => `
     color: white;
     background-color: ${colors[props.variant].main};
     border-color: ${colors[props.variant].main};
@@ -30,19 +31,21 @@ export const Button = styledComponentWithProps<{ variant: colorName }>(
     text-decoration: none;
   }
 
-  ${props => `
+  ${(props: Props) => `
     &:hover {
       background-color: ${colors[props.variant].contrast};
       border-color: ${colors[props.variant].contrast};
     }
   `};
 
-  &:focus, &.focus {
+  &:focus,
+  &.focus {
     outline: 0;
     box-shadow: 0 0 0 0.2rem rgba(0, 123, 255, 0.25);
   }
 
-  &.disabled, &:disabled {
+  &.disabled,
+  &:disabled {
     opacity: 0.65;
   }